back/requirements.txt
Flask==1.0.2
imageio==2.3.0
numpy==1.16.2
Pillow==6.2.0
requests==2.21.0
torch==1.3.1
torchvision==0.2.1
waitress==1.4.3
toml
easydict
tensorboardX
scipy
pprint
opencv-python-headless
